cloudfunctions-example/requirements.txt
# Function dependencies, for example:
# package>=version
opentelemetry-api==1.13.0
opentelemetry-sdk==1.13.0
opentelemetry-exporter-gcp-trace==1.3.0
opentelemetry-propagator-gcp==1.3.0
opentelemetry-instrumentation-flask==0.34b0